getting-started-with-Ollama/requirement.txt
langchain
python-dotenv
langchain-google-genai
ipykernel
beautifulsoup4
langchain_community
faiss-cpu
langchain_classic
streamlit